property_app/requirements.txt
asgiref==3.5.2
Django==3.2.13
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.49
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
